summarization_detector: split detect.py into a package, drop the profiler, review fixes
