+++ pyproject.toml
@@ -208,7 +208,6 @@ fixconf = "certora_autosetup.fixconf:main"
 
 # summarization_detector entry points
 detect-summaries = "summarization_detector.cli:main"
-detect-difficulty-profile = "summarization_detector.difficulty_profile:main"
 
 # Command-sandbox providers, resolved by composer.sandbox.config.SandboxConfig.resolve_provider.
 # Adding a provider = adding a line here (and re-running the editable install so the
