+++ summarization_detector/README.md
@@ -4,9 +4,8 @@ A standalone AutoProver tool that, from **one prover run**, ranks the functions 
 candidate list with *why* each is prover-hostile, *where* it can be summarized, and (for a curated
 public-library match) a suggested summary. It decides *what* to summarize, **not** *how*: the summarization
 strategy (per-function over-approximation, a whole-contract symbolic model, …) and the actual summary text
-are the downstream generator's job (curated summaries, the symbolic-model tool, or the CVL_GEN agent). A
-caller runs it after a slow/timeout (or sanity) run to summarize the expensive functions before paying for
-them.
+are the consumer's job. A caller runs it after a slow/timeout (or sanity) run to summarize the expensive
+functions before paying for them.
 
 It is self-contained: it reads the prover's difficulty report via its own `difficulty` module and uses
 `certora_autosetup` (the solc AST reader).
@@ -44,19 +43,3 @@ python -m summarization_detector --cut Router --ast .asts.json \
 
 `externalCallGraph.json` is emitted by the prover (EVMVerifier `ExternalCallSiteCollector`) at scene
 setup — after call resolution, before the per-rule optimize pass — and enables the reachability gate.
-
-## difficulty_profile — post-hoc: where did the prover time actually go?
-
-`detect.py` predicts what to summarize *before* the rules exist. `difficulty_profile.py` is its
-post-hoc counterpart: from a completed (esp. timed-out) run it reads the prover's own difficulty tree
-per slow rule and attributes the nonlinearity / path-count hotspots to source functions, classifying
-each as **cut** (a function of the contract under test → a generator's over-approx/precise model), **library**
-(inlined lib → library model), **external** (linked dependency → dependency model), or **cvl-model**
-(already summarized). The CUT and the scene's linked contracts are read from the run's treeViewStatus,
-so nothing is protocol-specific.
-
-    python -m summarization_detector.difficulty_profile <job-url-or-hash> [more jobs...] --min-minutes 20
-    python -m summarization_detector.difficulty_profile <hashes> --json      # for pipeline consumption
-
-Use it to decide, with evidence, whether a scene's timeouts are in dependency/library code (a dependency/library model
-dependency/library model helps) or in the contract under test itself (a per-function model / harness).
