+++ summarization_detector/tests/test_sources.py
@@ -5,18 +5,20 @@ import json
 import tempfile
 from pathlib import Path
 
+import pytest
+
 from summarization_detector.sources import (
     cut_from_conf, find_run_conf, find_external_call_graph)
 
 
-def test_cut_from_conf_reads_verify_then_parametric():
+def test_cut_from_conf_reads_verify_only():
     with tempfile.TemporaryDirectory() as d:
         c = Path(d) / "run.conf"
         c.write_text(json.dumps({"verify": "Router:certora/specs/sanity-Router.spec"}))
         assert cut_from_conf(c) == "Router"
-        c.write_text(json.dumps({"parametric_contracts": ["Vault", "Other"]}))   # no verify -> parametric
-        assert cut_from_conf(c) == "Vault"
-        c.write_text(json.dumps({"files": ["A.sol"]}))                            # neither -> ""
+        # `verify` is the authoritative field; without it we do NOT guess the CUT from `parametric_contracts`
+        # (which names all parametric targets, not the main contract) — the caller must pass `cut` explicitly.
+        c.write_text(json.dumps({"parametric_contracts": ["Vault", "Other"]}))
         assert cut_from_conf(c) == ""
 
 
@@ -29,6 +31,18 @@ def test_find_run_conf_prefers_canonical_and_skips_lib():
         (root / "lib" / "dep" / "run.conf").write_text("{}")                     # must be ignored
         found = find_run_conf(root)
         assert found == root / "inputs" / ".certora_sources" / "run.conf"
+
+
+def test_find_run_conf_raises_without_canonical_never_guesses():
+    # No canonical run.conf -> raise rather than silently pick some other .conf (a mock / sub-run / leftover
+    # would drive the whole run off the wrong scene). The candidates are listed to help, not chosen.
+    with tempfile.TemporaryDirectory() as d:
+        root = Path(d)
+        (root / "sub").mkdir()
+        (root / "sub" / "other.conf").write_text("{}")
+        with pytest.raises(FileNotFoundError) as ei:
+            find_run_conf(root)
+        assert "other.conf" in str(ei.value) and "canonical" in str(ei.value)
 
 
 def test_find_external_call_graph_optional():
